requirements.txt
# LangChain Core
langchain



# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

streamlit
uvicorn
pydantic
fastapi